pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Abiorh001 <abiolaadedayo1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.10"
llama-index = "^0.12.13"
python-dotenv = "^1.0.1"
cohere = "^5.13.11"
llama-index-postprocessor-cohere-rerank = "^0.3.0"
llama-index-retrievers-bm25 = "^0.5.2"
llama-index-vector-stores-pinecone = "^0.4.3"
graspologic = "^3.4.1"
numpy = "^1.26.4"
scipy = "1.12.0"
future = "^1.0.0"
llama-index-graph-stores-neo4j = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
